Inventory Core Implemented

+++ HomeAppliances/ShopManagement.Application.Contracts/Product/ProductSearchModel.cs
@@ -4,6 +4,6 @@
     {
         public string Name { get; set; }
         public string Code { get; set; }
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
     }
 }
